add_product_screen updated

+++ lib/models/product.dart
@@ -24,7 +24,7 @@ class ProductModel {
   String _description;
 
   double _rating;
-  int _price;
+  double _price;
   int _rates;
 
   bool _featured;
@@ -39,7 +39,7 @@ class ProductModel {
   String get image => _image;
 
   double get rating => _rating;
-  int get price => _price;
+  double get price => _price;
   bool get featured => _featured;
   int get rates => _rates;
 
@@ -54,9 +54,9 @@ class ProductModel {
     _description = snapshot.data()[DESCRIPTION];
     _id = snapshot.data()[ID];
     _featured = snapshot.data()[FEATURED];
-    _price = snapshot.data()[PRICE];
+    _price = snapshot.data()[PRICE].toDouble();
     _category = snapshot.data()[CATEGORY];
-    _rating = snapshot.data()[RATING];
+    _rating = snapshot.data()[RATING].toDouble();
     _rates = snapshot.data()[RATES];
     _name = snapshot.data()[NAME];
   }
